+++ Events.lua
@@ -64,10 +64,9 @@ function MyDungeonsBook:COMBAT_LOG_EVENT_UNFILTERED()
 		self:TrackBfASpecificCastDoneByPartyMembers(srcName, spellId);
 		self:TrackSLSpecificCastDoneByPartyMembers(srcName, spellId);
 		self:TrackAllCastsDoneByPartyMembers(srcName, srcGUID, spellId);
-		self:TrackBfAOwnCastDoneByPartyMembers(srcName, spellId, dstName);
-		self:TrackSLOwnCastDoneByPartyMembers(srcName, spellId, dstName);
 		self:TrackBfASpecificItemUsedByPartyMembers(srcName, spellId);
 		self:TrackSLSpecificItemUsedByPartyMembers(srcName, spellId);
+		self:TrackOwnCastDoneByPartyMembers(srcName, spellId, dstName);
 	end
 	if ((subEventPrefix:match("^SPELL") or subEventPrefix:match("^RANGE")) and subEventSuffix == "DAMAGE") then
 		local spellId, _, _, amount, overkill = select(12, CombatLogGetCurrentEventInfo());